Adds global ACL system, myaccess command
This commit adds functions getGroups and hasPriv, which form the meat of the
bot's access control system.  As their names imply, these functions do the
actual privilege checking.  This commit also adds the 'myaccess' command, which
allows a user to see what groups they are a member of in the global ACL.

Change-Id: Ibc335f89424754e18bcd3918e4b2cfbf20d96027