Add transmitLine(), add 'die' command.
This commit adds the transmitLine() function, which streamlines the process
of determining whether a message was sent in a channel or in a private message
(and automatically responds accordingly).

This commit also adds the 'die' command, which - when used by an authorized user
who also give the correct die password ($conf['diepass']), kills the bot
cleanly-ish.

Change-Id: I97a897192f139abf533c281efa7b988adeeff67a