Database prefixes, restart command, and ACL cmds
This commit adds proper support for database prefixes ($conf['dbPrefix']).

This commit also adds a restart command, which is the easiest way to make
configuration changes take effect.  It is used similarly to the die command,
requiring the diepass just like the die command as well as the 'die' privilege,
but it restarts the bot instead of outright killing it.

Finally, this commit adds the global ACL management commands, 'acl list', 'acl
add', 'acl setgroup', 'acl rm', and 'acl reload'.  Their names should be self-
explanatory, at least for the purposes of this commit message.  Some
reorganization of the bot's MySQL logic is also included for this.

Change-Id: I950aa44105b8d096522db41f875bc7e94e43e819