Add interwiki commands
This commit adds the 'interwiki' group of commands:

* 'interwiki list'
* 'interwiki add'
* 'interwiki rm'
* 'interwiki settarget'
* 'interwiki reload'

These commands act as the method of manipulating the bot's interwiki prefix
list.  In addition, to make things easier, the interwiki_prefix column in the
interwiki table has been made a unique key, and for better presentation, the
interwiki list is now sorted by interwiki prefix, not interwiki ID.

Bonus: In preparation for adding the channel table manipulation commands, the
channel_name column in the channel table has also been made a unique key.

Change-Id: I9dd859ae5ce91ec32c6fc73f31501ea96db4a7b9