Add SAPI check, argument parsing, --config option
This commit, firstly, adds a PHP SAPI check to ensure that the script is only
being run from the command line.  Note that this check may fail if the script
is being called with the PHP CGI binary from the command line - this is
intentional (the script's behavior is undefined when called using anything
other than the proper PHP CLI binary).

Next, this commit adds basic command line argument and option parsing, and adds
a --config (-c for short) option that allows for selection of a config file
other than the default.  The standard --, --help, and --version switches are
also added.  (Note that no non-option arguments are accepted by the script
currently.)

Change-Id: I710f52598d0b0bec9a1bd9201913c1727db6dacf