Add STATUSMSG char detection, update .gitignore
This commit adds STATUSMSG detection to the bot, which allows it to detect which
single-character "status messages" are used by the network.  Here are the most
common ones:

~ = Owner
& = Protected (AKA Admin)
@ = Operator
% = Half-operator
+ = Voice

The bot assumes that only '@' and '+' are supported, until it receives an IRC
numeric 005 with a STATUSMSG capability listing, at which point it will use that
to determine what status messages are supported by the network.  Since this is
part of the IRC standard (de facto, at least), this should work on all networks.

(It also conveniently turns out that the 005 numeric gives the relative "power"
or "ranking", if you will, of the various status messages by listing, from left-
to-right, the highest ranking one first and proceeding in descending order.)

This paves the way for channel management access controls (which will be based
on a user's status in a channel instead of an ACL like the Global ACL system),
and once the channel management system is implemented, we can implement actual
wikilink expansion.  Yay!

This commit also updates .gitignore to ignore config files with names other than
"lizardlinky.conf.inc".

Change-Id: Iaaf461a4cda8519d400772b7f5ce582c2c2110d2